requirements.txt
scrapy>=2.8.0
pymysql>=1.0.2
redis>=4.5.0
python-dotenv>=1.0.0
